UserService finished

+++ tec/cod/thedrophunter/src/main/java/es/theDropHunter/StartUp/repositories/UserRepository.java
@@ -1,6 +1,6 @@
 package es.theDropHunter.StartUp.repositories;
 
-import java.util.List;
+
 import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,7 +9,7 @@ import es.theDropHunter.StartUp.entities.User;
 
 public interface UserRepository extends CrudRepository<User, Long>{
 
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 
     
 }
